ipcalc/requirements.txt
cloudshell-shell-core>=2.0.0,<2.1.0
cloudshell_automation_api
natsort
ipcalc